Refactor project structure to move API into services directory
Moved API-related code, including main.go, routes, and middleware, from the api directory to services/api. This reorganization improves the project's modular structure and aligns related services better within the directory hierarchy.
